+++ src/hkim0331/util/bench.clj
@@ -1,4 +1,16 @@
-(ns util.bench)
+(ns hkim0331.util.bench
+  (:require [criterium.core :refer [with-progress-reporting quick-bench]]))
+
+#_(defmacro quick [expr]
+    `(with-progress-reporting
+       (quick-bench ~expr :verbose)))
+
+(defmacro with-progress [expr]
+  `(with-progress-reporting
+     (quick-bench ~expr))) ; without :verbose
+
+(defmacro quick [expr]
+  `(quick-bench ~expr)) ; without progress reporting)
 
 (let [time*
       (fn [^long duration-in-ms f]
@@ -44,17 +56,19 @@
     (format "Used: %.0f/%.0f MB (%.0f%%), free: %.0f MB" used total (/ used total 0.01)
             (/ (.freeMemory (Runtime/getRuntime)) 1e6))))
 
-(defmacro debugging-tools []
-  '(do
-     (require '[clj-async-profiler.core :as prof])
-     (require '[clj-java-decompiler.core :refer [decompile]])
-     (require '[clj-memory-meter.core :as mm])
+; what is this?
+;
+; (defmacro debugging-tools []
+;   '(do
+;      (require '[clj-async-profiler.core :as prof])
+;      (require '[clj-java-decompiler.core :refer [decompile]])
+;      (require '[clj-memory-meter.core :as mm])
 
-     (.refer *ns* 'time+ #'user/time+)
-     (.refer *ns* 'heap #'user/heap)))
+;      (.refer *ns* 'time+ #'user/time+)
+;      (.refer *ns* 'heap #'user/heap)))
 
-(comment
-  (debugging-tools)
-  ; Execution error (FileNotFoundException) at user/eval11131 (REPL:55).
-  ; Could not locate clj_async_profiler/core__init.class, clj_async_profiler/core.clj or clj_async_profiler/core.cljc on classpath. Please check that namespaces with dashes use underscores in the Clojure file name.
-  :rcf)
+; (comment
+;   (debugging-tools)
+;   ; Execution error (FileNotFoundException) at user/eval11131 (REPL:55).
+;   ; Could not locate clj_async_profiler/core__init.class, clj_async_profiler/core.clj or clj_async_profiler/core.cljc on classpath. Please check that namespaces with dashes use underscores in the Clojure file name.
+;   :rcf)