+++ CHANGELOG.md
@@ -1,13 +1,40 @@
 # Change Log
 
+Learn how to make/use github libraries.
+
 Under construction.
 
 ## [Unreleased]
 
-- learn how to use.
 - more utils.
 - git/sha is determined at push time, master branch to GITHUB.
-  how to reflect the :git/sha value into updated README.md?
+  how to reflect the :git/sha value into README.md?
+- criterium: hkim0331.util.bench requires criterium.
+
+## v0.4-SNAPSHOT
+
+- core/reverse-all
+- core/flatten-all
+
+## v0.4.65 / 2025-02-09
+
+- core/prime-pi
+
+## v0.4.59 / 2025-02-08
+
+- `hkim0331.util` namespace.
+
+## v0.3.53 / 2025-02-08
+
+- debugged divisors
+- included test dir in `just dev`.
+- just dev starts interactive REPL, too.
+- improved cart with `last` and `butlast`.
+
+## v2.0.33 / 2025-02-07
+
+- improved bench.clj
+- divisors using factor-integer. 5 times faster.
 
 ## v0.2.1-SNAPSHOT
 
@@ -19,13 +46,11 @@ Under construction.
 - changed name of receipies. 'dev', 'dev-container'.
 - bump-version.sh
 
-
 ## v0.1.0 - 2025-02-04
 
 - primes
 - fold-l, fold-r
-- tarai, tarai-memo, tarai-lazy
-
+- tarai, tarai-memo, tarai-lazyd
 
 ## [v0.0.3] - 2025-02-04